src/cloud-function/requirements.txt
google-cloud-aiplatform[pipelines]==1.12.1
google-cloud-storage==2.3.0